Major bug fixes and improvements
Added asynchronous loading when performing a search. Now only the list of people found in the search is loaded. The views containing relations and details of a person are not loaded until they are opened.
Fixed typo in Civil Status “Registreret partner”.
Fixed value error in world addresses and danish addresses views.
People living on same address are now accurate.